requirements.txt
Flask
python-dotenv
gunicorn
flask_cors
six
transformers
sentencepiece
protobuf
torch